Add new plugin 'Chris' Toolbar Publish Button'
The added plugin has the Toolbar Publish Button that retains the scrollbar position after saving to improve user experience in the WordPress admin area. It includes options to keep the scrollbar position after saving, choose which buttons to show on the admin bar, move buttons to the right side of the Toolbar, and set a background color for the buttons. It's well-integrated with the Advanced Custom Fields and helps to reduce excessive scrolling.